+++ src/devices/mod.rs
@@ -1,9 +1,3 @@
-#[derive(Debug, Clone, Copy)]
-enum OperatingMode {
-    OneShot,
-    Continuous,
-}
-
 mod common;
 mod features;
 mod mode;