+++ src/types.rs
@@ -1,11 +1,9 @@
 //! Type definitions.
 
-use core::marker::PhantomData;
-
-use crate::{private, ChannelSelection};
+use crate::Config;
 
 /// Errors in this crate
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 pub enum Error<E> {
     /// I²C bus error
     I2C(E),
@@ -13,27 +11,7 @@ pub enum Error<E> {
     InvalidInputData,
 }
 
-/// Error type for mode changes.
-///
-/// This allows to retrieve the unchanged device in case of an error.
-pub enum ModeChangeError<E, DEV> {
-    /// I²C bus error while changing mode.
-    ///
-    /// `E` is the error that happened.
-    /// `DEV` is the device with the mode unchanged.
-    I2C(E, DEV),
-}
-
-/// Mode marker types
-pub mod mode {
-    /// One-shot operating mode / power-down state (default)
-    pub struct OneShot(());
-
-    /// Continuous conversion mode
-    pub struct Continuous(());
-}
-
-/// Data rate for ADS1013, ADS1014, ADS1015
+/// Data rate for ADS101x.
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Ord, PartialOrd, Hash)]
 pub enum DataRate12Bit {
     /// 128 SPS
@@ -53,7 +31,42 @@ pub enum DataRate12Bit {
     Sps3300,
 }
 
-/// Data rate for ADS1113, ADS1114, ADS1115
+impl DataRate12Bit {
+    pub(crate) fn configure(self, cfg: Config) -> Config {
+        match self {
+            Self::Sps128 => cfg
+                .difference(Config::DR2)
+                .difference(Config::DR1)
+                .difference(Config::DR0),
+            Self::Sps250 => cfg
+                .difference(Config::DR2)
+                .difference(Config::DR1)
+                .union(Config::DR0),
+            Self::Sps490 => cfg
+                .difference(Config::DR2)
+                .union(Config::DR1)
+                .difference(Config::DR0),
+            Self::Sps920 => cfg
+                .difference(Config::DR2)
+                .union(Config::DR1)
+                .union(Config::DR0),
+            Self::Sps1600 => cfg
+                .union(Config::DR2)
+                .difference(Config::DR1)
+                .difference(Config::DR0),
+            Self::Sps2400 => cfg
+                .union(Config::DR2)
+                .difference(Config::DR1)
+                .union(Config::DR0),
+            Self::Sps3300 => cfg
+                .union(Config::DR2)
+                .union(Config::DR1)
+                .difference(Config::DR0),
+        }
+    }
+}
+
+/// Data rate for ADS111x.
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Ord, PartialOrd, Hash)]
 pub enum DataRate16Bit {
     /// 8 SPS
@@ -75,7 +88,43 @@ pub enum DataRate16Bit {
     Sps860,
 }
 
-/// Comparator mode (only for ADS1x14, ADS1x15)
+impl DataRate16Bit {
+    pub(crate) fn configure(self, cfg: Config) -> Config {
+        match self {
+            Self::Sps8 => cfg
+                .difference(Config::DR2)
+                .difference(Config::DR1)
+                .difference(Config::DR0),
+            Self::Sps16 => cfg
+                .difference(Config::DR2)
+                .difference(Config::DR1)
+                .union(Config::DR0),
+            Self::Sps32 => cfg
+                .difference(Config::DR2)
+                .union(Config::DR1)
+                .difference(Config::DR0),
+            Self::Sps64 => cfg
+                .difference(Config::DR2)
+                .union(Config::DR1)
+                .union(Config::DR0),
+            Self::Sps128 => cfg
+                .union(Config::DR2)
+                .difference(Config::DR1)
+                .difference(Config::DR0),
+            Self::Sps250 => cfg
+                .union(Config::DR2)
+                .difference(Config::DR1)
+                .union(Config::DR0),
+            Self::Sps475 => cfg
+                .union(Config::DR2)
+                .union(Config::DR1)
+                .difference(Config::DR0),
+            Self::Sps860 => cfg.union(Config::DR2).union(Config::DR1).union(Config::DR0),
+        }
+    }
+}
+
+/// Comparator mode (only for ADS1x14, ADS1x15).
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Ord, PartialOrd, Hash)]
 pub enum ComparatorMode {
     #[default]
@@ -94,7 +143,7 @@ pub enum ComparatorMode {
     Window,
 }
 
-/// Comparator polarity (only for ADS1x14, ADS1x15)
+/// Comparator polarity (only for ADS1x14, ADS1x15).
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Ord, PartialOrd, Hash)]
 pub enum ComparatorPolarity {
     #[default]
@@ -104,7 +153,7 @@ pub enum ComparatorPolarity {
     ActiveHigh,
 }
 
-/// Comparator polarity (only for ADS1x14, ADS1x15)
+/// Comparator latching (only for ADS1x14, ADS1x15).
 ///
 /// Select whether the ALERT/RDY pin latches after being asserted or clears
 /// after conversions are within the margin of the upper and lower
@@ -125,7 +174,7 @@ pub enum ComparatorLatching {
     Latching,
 }
 
-/// Comparator alert queue (only for ADS1x14, ADS1x15)
+/// Comparator alert queue (only for ADS1x14, ADS1x15).
 ///
 /// The default state of the comparator is deactivated. It can be activated by setting
 /// the comparator queue.
@@ -140,156 +189,103 @@ pub enum ComparatorQueue {
     Four,
 }
 
-/// Full-scale range configuration for the programmable gain amplifier (PGA) (only for ADS1x14, ADS1x15)
+/// Full-scale range configuration for the programmable gain amplifier (PGA) (only for ADS1x14, ADS1x15).
 ///
 /// This sets the input voltage measurable range.
 /// The FSR is fixed at ±2.048 V in the ADS1x13.
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Ord, PartialOrd, Hash)]
 #[allow(non_camel_case_types)]
 pub enum FullScaleRange {
-    /// The measurable range is ±6.144V.
+    /// ±6.144V
     Within6_144V,
-    /// The measurable range is ±4.096V.
+    /// ±4.096V
     Within4_096V,
+    /// ±2.048V (default)
     #[default]
-    /// The measurable range is ±2.048V. (default)
     Within2_048V,
-    /// The measurable range is ±1.024V.
+    /// ±1.024V
     Within1_024V,
-    /// The measurable range is ±0.512V.
+    /// ±0.512V
     Within0_512V,
-    /// The measurable range is ±0.256V.
+    /// ±0.256V
     Within0_256V,
 }
 
-/// Possible slave addresses
+impl FullScaleRange {
+    pub(crate) fn configure(self, config: Config) -> Config {
+        match self {
+            Self::Within6_144V => config
+                .difference(Config::PGA2)
+                .difference(Config::PGA1)
+                .difference(Config::PGA0),
+            Self::Within4_096V => config
+                .difference(Config::PGA2)
+                .difference(Config::PGA1)
+                .union(Config::PGA0),
+            Self::Within2_048V => config
+                .difference(Config::PGA2)
+                .union(Config::PGA1)
+                .difference(Config::PGA0),
+            Self::Within1_024V => config
+                .difference(Config::PGA2)
+                .union(Config::PGA1)
+                .union(Config::PGA0),
+            Self::Within0_512V => config
+                .union(Config::PGA2)
+                .difference(Config::PGA1)
+                .difference(Config::PGA0),
+            Self::Within0_256V => config
+                .union(Config::PGA2)
+                .difference(Config::PGA1)
+                .union(Config::PGA0),
+        }
+    }
+}
+
+/// A slave address.
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Ord, PartialOrd, Hash)]
 pub enum SlaveAddr {
+    /// Address when the ADDR pin is connected to GND. (default)
     #[default]
-    /// Default slave address
-    Default,
-    /// Alternative slave address using the provided values
-    /// for the last two bits (A1, A0)
-    Alternative(bool, bool),
+    Gnd,
+    /// Address when the ADDR pin is connected to VDD.
+    Vdd,
+    /// Address when the ADDR pin is connected to SDA.
+    ///
+    /// If SDA is used as the device address, hold the SDA line low for at
+    /// least 100 ns after the SCL line goes low to make sure the device
+    /// decodes the address correctly during I²C communication.
+    Sda,
+    /// Address when the ADDR pin is connected to SCL.
+    Scl,
 }
 
 impl SlaveAddr {
-    pub(crate) fn addr(self, default: u8) -> u8 {
+    pub(crate) const fn bits(self) -> u8 {
         match self {
-            SlaveAddr::Default => default,
-            SlaveAddr::Alternative(a1, a0) => default | ((a1 as u8) << 1) | a0 as u8,
+            SlaveAddr::Gnd => 0b100_1000,
+            SlaveAddr::Vdd => 0b100_1001,
+            SlaveAddr::Sda => 0b100_1010,
+            SlaveAddr::Scl => 0b100_1011,
         }
     }
-
-    /// Create `SlaveAddr` instance corresponding to the address
-    /// effective when connecting the pin `ADDR` to GND (0x48).
-    ///
-    /// See [Table 4 in the datasheet](https://www.ti.com/lit/ds/symlink/ads1115.pdf#%5B%7B%22num%22%3A716%2C%22gen%22%3A0%7D%2C%7B%22name%22%3A%22XYZ%22%7D%2C0%2C602.2%2C0%5D).
-    pub fn new_gnd() -> Self {
-        SlaveAddr::default()
-    }
-
-    /// Create `SlaveAddr` instance corresponding to the address
-    /// effective when connecting the pin `ADDR` to VDD (0x49).
-    ///
-    /// See [Table 4 in the datasheet](https://www.ti.com/lit/ds/symlink/ads1115.pdf#%5B%7B%22num%22%3A716%2C%22gen%22%3A0%7D%2C%7B%22name%22%3A%22XYZ%22%7D%2C0%2C602.2%2C0%5D).
-    pub fn new_vdd() -> Self {
-        SlaveAddr::Alternative(false, true)
-    }
-
-    /// Create `SlaveAddr` instance corresponding to the address
-    /// effective when connecting the pin `ADDR` to SDA (0x4A).
-    ///
-    /// See [Table 4 in the datasheet](https://www.ti.com/lit/ds/symlink/ads1115.pdf#%5B%7B%22num%22%3A716%2C%22gen%22%3A0%7D%2C%7B%22name%22%3A%22XYZ%22%7D%2C0%2C602.2%2C0%5D).
-    pub fn new_sda() -> Self {
-        SlaveAddr::Alternative(true, false)
-    }
-
-    /// Create `SlaveAddr` instance corresponding to the address
-    /// effective when connecting the pin `ADDR` to SCL (0x4B).
-    ///
-    /// See [Table 4 in the datasheet](https://www.ti.com/lit/ds/symlink/ads1115.pdf#%5B%7B%22num%22%3A716%2C%22gen%22%3A0%7D%2C%7B%22name%22%3A%22XYZ%22%7D%2C0%2C602.2%2C0%5D).
-    pub fn new_scl() -> Self {
-        SlaveAddr::Alternative(true, true)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct Config {
-    pub(crate) bits: u16,
-}
-
-impl Config {
-    pub(crate) fn is_high(&self, mask: u16) -> bool {
-        (self.bits & mask) != 0
-    }
-
-    pub(crate) fn with_high(&self, mask: u16) -> Self {
-        Config {
-            bits: self.bits | mask,
-        }
-    }
-
-    pub(crate) fn with_low(&self, mask: u16) -> Self {
-        Config {
-            bits: self.bits & !mask,
-        }
-    }
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Config { bits: 0x8583 }
-    }
-}
-
-/// ADS1x1x ADC driver
-#[derive(Debug, Default)]
-pub struct Ads1x1x<I2C, IC, CONV, MODE> {
-    pub(crate) i2c: I2C,
-    pub(crate) address: u8,
-    pub(crate) config: Config,
-    pub(crate) fsr: FullScaleRange,
-    pub(crate) a_conversion_was_started: bool,
-    pub(crate) _conv: PhantomData<CONV>,
-    pub(crate) _ic: PhantomData<IC>,
-    pub(crate) _mode: PhantomData<MODE>,
-}
-
-/// Multi channel One-shot ADC
-pub trait DynamicOneShot: private::Sealed {
-    /// Error type
-    type Error;
-
-    /// Read a measurement
-    fn read(&mut self, channel: ChannelSelection) -> nb::Result<i16, Self::Error>;
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::DEVICE_BASE_ADDRESS as ADDR;
     use crate::{FullScaleRange, SlaveAddr};
 
     #[test]
-    fn can_get_default_address() {
-        let addr = SlaveAddr::default();
-        assert_eq!(ADDR, addr.addr(ADDR));
+    fn slave_addr_default() {
+        assert_eq!(0b100_1000, SlaveAddr::default().bits());
     }
 
     #[test]
-    fn can_generate_alternative_addresses() {
-        assert_eq!(0b100_1000, SlaveAddr::Alternative(false, false).addr(ADDR));
-        assert_eq!(0b100_1001, SlaveAddr::Alternative(false, true).addr(ADDR));
-        assert_eq!(0b100_1010, SlaveAddr::Alternative(true, false).addr(ADDR));
-        assert_eq!(0b100_1011, SlaveAddr::Alternative(true, true).addr(ADDR));
-    }
-
-    #[test]
-    fn can_generate_alternative_addresses_using_helper_constructors() {
-        assert_eq!(0b100_1000, SlaveAddr::new_gnd().addr(ADDR));
-        assert_eq!(0b100_1001, SlaveAddr::new_vdd().addr(ADDR));
-        assert_eq!(0b100_1010, SlaveAddr::new_sda().addr(ADDR));
-        assert_eq!(0b100_1011, SlaveAddr::new_scl().addr(ADDR));
+    fn slave_addr_bits() {
+        assert_eq!(0b100_1000, SlaveAddr::Gnd.bits());
+        assert_eq!(0b100_1001, SlaveAddr::Vdd.bits());
+        assert_eq!(0b100_1010, SlaveAddr::Sda.bits());
+        assert_eq!(0b100_1011, SlaveAddr::Scl.bits());
     }
 
     #[test]