+++ tests/tier1.rs
@@ -18,22 +18,22 @@ macro_rules! measure_tests {
 
                 #[test]
                 fn read_if_measurement_in_progress() {
-                    let config = Config::default().with_low(BF::OS);
+                    let config = Config::default().difference(BF::OS);
                     let transactions = [I2cTrans::write_read(
                         DEV_ADDR,
                         vec![Register::CONFIG],
                         vec![config.msb(), config.lsb()],
                     )];
-                    let mut dev = $create(&transactions);
-                    assert_would_block!(dev.read(channel::DifferentialA0A1));
-                    $destroy(dev);
+                    let mut adc = $create(&transactions);
+                    assert_would_block!(adc.read(channel::DifferentialA0A1));
+                    $destroy(adc);
                 }
             }
 
             #[test]
             fn can_measure() {
                 let default_config = Config::default();
-                let config_with_os = Config::default().with_high(BF::OS);
+                let config_with_os = Config::default().union(BF::OS);
                 let transactions = [
                     I2cTrans::write_read(
                         DEV_ADDR,
@@ -51,24 +51,24 @@ macro_rules! measure_tests {
                     ),
                     I2cTrans::write_read(DEV_ADDR, vec![Register::CONVERSION], vec![0x80, 0x00]),
                 ];
-                let mut dev = $create(&transactions);
-                let measurement = block!(dev.read(channel::DifferentialA0A1)).unwrap();
+                let mut adc = $create(&transactions);
+                let measurement = block!(adc.read(channel::DifferentialA0A1)).unwrap();
                 assert_eq!($expected, measurement);
-                $destroy(dev);
+                $destroy(adc);
             }
 
             #[test]
             fn can_measure_continuous() {
-                let config = Config::default().with_low(BF::OP_MODE);
+                let config = Config::default().difference(BF::OP_MODE);
                 let transactions = [
                     I2cTrans::write(DEV_ADDR, vec![Register::CONFIG, config.msb(), config.lsb()]),
                     I2cTrans::write_read(DEV_ADDR, vec![Register::CONVERSION], vec![0x80, 0x00]),
                 ];
-                let dev = $create(&transactions);
-                let mut dev = dev.into_continuous().ok().unwrap();
-                let measurement = dev.read().unwrap();
+                let adc = $create(&transactions);
+                let mut adc = adc.into_continuous().ok().unwrap();
+                let measurement = adc.read().unwrap();
                 assert_eq!($expected, measurement);
-                $destroy(dev);
+                $destroy(adc);
             }
         }
     };
@@ -88,9 +88,9 @@ mod data_rate_12bit {
                     DEV_ADDR,
                     vec![Register::CONFIG, $config.msb(), $config.lsb()],
                 )];
-                let mut dev = new_ads1013(&transactions);
-                dev.set_data_rate(DataRate12Bit::$variant).unwrap();
-                destroy_ads1013(dev);
+                let mut adc = new_ads1013(&transactions);
+                adc.set_data_rate(DataRate12Bit::$variant).unwrap();
+                destroy_ads1013(adc);
             }
         };
     }
@@ -99,57 +99,57 @@ mod data_rate_12bit {
         sps128,
         Sps128,
         Config::default()
-            .with_low(BF::DR2)
-            .with_low(BF::DR1)
-            .with_low(BF::DR0)
+            .difference(BF::DR2)
+            .difference(BF::DR1)
+            .difference(BF::DR0)
     );
     test!(
         sps250,
         Sps250,
         Config::default()
-            .with_low(BF::DR2)
-            .with_low(BF::DR1)
-            .with_high(BF::DR0)
+            .difference(BF::DR2)
+            .difference(BF::DR1)
+            .union(BF::DR0)
     );
     test!(
         sps490,
         Sps490,
         Config::default()
-            .with_low(BF::DR2)
-            .with_high(BF::DR1)
-            .with_low(BF::DR0)
+            .difference(BF::DR2)
+            .union(BF::DR1)
+            .difference(BF::DR0)
     );
     test!(
         sps920,
         Sps920,
         Config::default()
-            .with_low(BF::DR2)
-            .with_high(BF::DR1)
-            .with_high(BF::DR0)
+            .difference(BF::DR2)
+            .union(BF::DR1)
+            .union(BF::DR0)
     );
     test!(
         sps1600,
         Sps1600,
         Config::default()
-            .with_high(BF::DR2)
-            .with_low(BF::DR1)
-            .with_low(BF::DR0)
+            .union(BF::DR2)
+            .difference(BF::DR1)
+            .difference(BF::DR0)
     );
     test!(
         sps2400,
         Sps2400,
         Config::default()
-            .with_high(BF::DR2)
-            .with_low(BF::DR1)
-            .with_high(BF::DR0)
+            .union(BF::DR2)
+            .difference(BF::DR1)
+            .union(BF::DR0)
     );
     test!(
         sps3300,
         Sps3300,
         Config::default()
-            .with_high(BF::DR2)
-            .with_high(BF::DR1)
-            .with_low(BF::DR0)
+            .union(BF::DR2)
+            .union(BF::DR1)
+            .difference(BF::DR0)
     );
 }
 
@@ -164,9 +164,9 @@ mod data_rate_16bit {
                     DEV_ADDR,
                     vec![Register::CONFIG, $config.msb(), $config.lsb()],
                 )];
-                let mut dev = new_ads1113(&transactions);
-                dev.set_data_rate(DataRate16Bit::$variant).unwrap();
-                destroy_ads1113(dev);
+                let mut adc = new_ads1113(&transactions);
+                adc.set_data_rate(DataRate16Bit::$variant).unwrap();
+                destroy_ads1113(adc);
             }
         };
     }
@@ -175,109 +175,109 @@ mod data_rate_16bit {
         sps8,
         Sps8,
         Config::default()
-            .with_low(BF::DR2)
-            .with_low(BF::DR1)
-            .with_low(BF::DR0)
+            .difference(BF::DR2)
+            .difference(BF::DR1)
+            .difference(BF::DR0)
     );
     test!(
         sps16,
         Sps16,
         Config::default()
-            .with_low(BF::DR2)
-            .with_low(BF::DR1)
-            .with_high(BF::DR0)
+            .difference(BF::DR2)
+            .difference(BF::DR1)
+            .union(BF::DR0)
     );
     test!(
         sps32,
         Sps32,
         Config::default()
-            .with_low(BF::DR2)
-            .with_high(BF::DR1)
-            .with_low(BF::DR0)
+            .difference(BF::DR2)
+            .union(BF::DR1)
+            .difference(BF::DR0)
     );
     test!(
         sps64,
         Sps64,
         Config::default()
-            .with_low(BF::DR2)
-            .with_high(BF::DR1)
-            .with_high(BF::DR0)
+            .difference(BF::DR2)
+            .union(BF::DR1)
+            .union(BF::DR0)
     );
     test!(
         sps128,
         Sps128,
         Config::default()
-            .with_high(BF::DR2)
-            .with_low(BF::DR1)
-            .with_low(BF::DR0)
+            .union(BF::DR2)
+            .difference(BF::DR1)
+            .difference(BF::DR0)
     );
     test!(
         sps250,
         Sps250,
         Config::default()
-            .with_high(BF::DR2)
-            .with_low(BF::DR1)
-            .with_high(BF::DR0)
+            .union(BF::DR2)
+            .difference(BF::DR1)
+            .union(BF::DR0)
     );
     test!(
         sps475,
         Sps475,
         Config::default()
-            .with_high(BF::DR2)
-            .with_high(BF::DR1)
-            .with_low(BF::DR0)
+            .union(BF::DR2)
+            .union(BF::DR1)
+            .difference(BF::DR0)
     );
     test!(
         sps860,
         Sps860,
         Config::default()
-            .with_high(BF::DR2)
-            .with_high(BF::DR1)
-            .with_high(BF::DR0)
+            .union(BF::DR2)
+            .union(BF::DR1)
+            .union(BF::DR0)
     );
 }
 
 #[test]
 fn can_read_measurement_in_progress() {
-    let config_os = Config::default().with_low(BF::OS);
+    let config_os = Config::default().difference(BF::OS);
     let transactions = [I2cTrans::write_read(
         DEV_ADDR,
         vec![Register::CONFIG],
         vec![config_os.msb(), config_os.lsb()],
     )];
-    let mut dev = new_ads1013(&transactions);
-    assert!(dev.is_measurement_in_progress().unwrap());
-    destroy_ads1013(dev);
+    let mut adc = new_ads1013(&transactions);
+    assert!(adc.is_measurement_in_progress().unwrap());
+    destroy_ads1013(adc);
 }
 
 #[test]
 fn can_read_measurement_not_in_progress() {
-    let config_os = Config::default().with_high(BF::OS);
+    let config_os = Config::default().union(BF::OS);
     let transactions = [I2cTrans::write_read(
         DEV_ADDR,
         vec![Register::CONFIG],
         vec![config_os.msb(), config_os.lsb()],
     )];
-    let mut dev = new_ads1013(&transactions);
-    assert!(!dev.is_measurement_in_progress().unwrap());
-    destroy_ads1013(dev);
+    let mut adc = new_ads1013(&transactions);
+    assert!(!adc.is_measurement_in_progress().unwrap());
+    destroy_ads1013(adc);
 }
 
 #[test]
 fn can_convert_to_continuous() {
-    let config = Config::default().with_low(BF::OP_MODE);
+    let config = Config::default().difference(BF::OP_MODE);
     let transactions = [I2cTrans::write(
         DEV_ADDR,
         vec![Register::CONFIG, config.msb(), config.lsb()],
     )];
-    let dev = new_ads1013(&transactions);
-    let dev = dev.into_continuous().ok().unwrap();
-    destroy_ads1013(dev);
+    let adc = new_ads1013(&transactions);
+    let adc = adc.into_continuous().ok().unwrap();
+    destroy_ads1013(adc);
 }
 
 #[test]
 fn can_convert_to_one_shot() {
-    let config_cont = Config::default().with_low(BF::OP_MODE);
+    let config_cont = Config::default().difference(BF::OP_MODE);
     let config_os = Config::default();
     let transactions = [
         I2cTrans::write(
@@ -289,8 +289,8 @@ fn can_convert_to_one_shot() {
             vec![Register::CONFIG, config_os.msb(), config_os.lsb()],
         ),
     ];
-    let dev = new_ads1013(&transactions);
-    let dev = dev.into_continuous().ok().unwrap();
-    let dev = dev.into_one_shot().ok().unwrap();
-    destroy_ads1013(dev);
+    let adc = new_ads1013(&transactions);
+    let adc = adc.into_continuous().ok().unwrap();
+    let adc = adc.into_one_shot().ok().unwrap();
+    destroy_ads1013(adc);
 }